Check user permission to edit/update/delete post

+++ resources/views/groups/show.blade.php
@@ -44,8 +44,10 @@
 	  				<td>{{ $post->user->email }}</td>
 	  				<td>{{ $post->created_at }}</td>
 	  				<td>
-	  					<a href="{{ route('groups.posts.edit', compact('group', 'post')) }}" class="btn btn-sm btn-default">編輯</a>
-	  					<a href="{{ route('groups.posts.destroy', compact('group', 'post')) }}" class="btn btn-sm btn-default" data-method="delete" data-confirm="Are you sure?" data-token="{{csrf_token()}}">刪除</a>
+	  					@can('update', $post)
+		  					<a href="{{ route('groups.posts.edit', compact('group', 'post')) }}" class="btn btn-sm btn-default">編輯</a>
+		  					<a href="{{ route('groups.posts.destroy', compact('group', 'post')) }}" class="btn btn-sm btn-default" data-method="delete" data-confirm="Are you sure?" data-token="{{csrf_token()}}">刪除</a>
+		  				@endcan
 	  				</td>
 	  			</tr>
 	  		@endforeach
